code/data-collectors/requirements.txt
mongo
feedparser
